Created init.sql
Created the inital file for the postgreSQL database. Added a table to store information about the schools, the leads, any and all AI queries, information to enrich leads, each job the redis data store has, the users using this application, an error log, the history of imported batches, an audit log, and an index.